finally connected admin page
